cv_storage: extract models.py
